Pull old FormListDownloader out of new component to make replacing simpler
