+++ collect_app/src/main/java/org/odk/collect/android/tasks/DownloadFormListTask.java
@@ -15,11 +15,12 @@
 package org.odk.collect.android.tasks;
 
 import android.os.AsyncTask;
+
 import androidx.annotation.Nullable;
 
-import org.odk.collect.android.formmanagement.ServerFormsDetailsFetcher;
 import org.odk.collect.android.listeners.FormListDownloaderListener;
 import org.odk.collect.android.logic.FormDetails;
+import org.odk.collect.android.utilities.FormListDownloader;
 
 import java.util.HashMap;
 
@@ -33,20 +34,20 @@ import java.util.HashMap;
  */
 public class DownloadFormListTask extends AsyncTask<Void, String, HashMap<String, FormDetails>> {
 
-    private final ServerFormsDetailsFetcher serverFormsDetailsFetcher;
+    private final FormListDownloader formListDownloader;
 
     private FormListDownloaderListener stateListener;
     private String url;
     private String username;
     private String password;
 
-    public DownloadFormListTask(ServerFormsDetailsFetcher serverFormsDetailsFetcher) {
-        this.serverFormsDetailsFetcher = serverFormsDetailsFetcher;
+    public DownloadFormListTask(FormListDownloader formListDownloader) {
+        this.formListDownloader = formListDownloader;
     }
 
     @Override
     protected HashMap<String, FormDetails> doInBackground(Void... values) {
-        return serverFormsDetailsFetcher.downloadFormList(url, username, password, false);
+        return formListDownloader.downloadFormList(url, username, password, false);
     }
 
     @Override
